Removed obsolete RxPy code
