--- dronecode_sdk/plugins/__init__.py
@@ -1,3 +1,0 @@
-# -*- coding: utf-8 -*-
-from .action import Action
-from .telemetry import Telemetry
